+++ extensions/pilot-ynwr/src/calendar.tsx
@@ -1,3 +1,4 @@
+import UseOAuth from "./fetch/useOAuth";
 import useDBLinkHook from "./hooks/DBLinkHook";
 import SelectDBsForm from "./views/forms/SelectDBsForm";
 import CalendarView from "./views/lists/CalendarView";
@@ -5,5 +6,7 @@ import CalendarView from "./views/lists/CalendarView";
 export default function Command() {
   const { linked } = useDBLinkHook();
 
-  return linked ? <CalendarView /> : <SelectDBsForm />;
+  const { notion } = UseOAuth();
+
+  return linked ? <CalendarView notion={notion} /> : <SelectDBsForm notion={notion} />;
 }