+++ extensions/pilot-ynwr/src/journal.tsx
@@ -1,3 +1,4 @@
+import UseOAuth from "./fetch/useOAuth";
 import useDBLinkHook from "./hooks/DBLinkHook";
 import SelectDBsForm from "./views/forms/SelectDBsForm";
 import JournalView from "./views/lists/JournalView";
@@ -5,5 +6,7 @@ import JournalView from "./views/lists/JournalView";
 export default function Command() {
   const { linked } = useDBLinkHook();
 
-  return linked ? <JournalView /> : <SelectDBsForm />;
+  const { notion } = UseOAuth();
+
+  return linked ? <JournalView notion={notion} /> : <SelectDBsForm notion={notion} />;
 }