+++ extensions/pilot-ynwr/src/queriesFunctions/ProjectQueries.tsx
@@ -4,14 +4,14 @@ import {
   QueryDatabaseParameters,
   UpdatePageParameters,
 } from "@notionhq/client/build/src/api-endpoints";
-import { Cache, getPreferenceValues, showToast, Toast } from "@raycast/api";
-import { Pref, Project } from "../interfaces/interfaceItems";
+import { Cache, LocalStorage, showToast, Toast } from "@raycast/api";
+import { Project } from "../interfaces/interfaceItems";
 import { getAPIError } from "../tools/generalTools";
 import { getTimers } from "../fetch/ExportFunctions";
 import { FetchedTodos, FetchEvents, FetchJournals, FetchKeystones } from "../fetch/FetchFunctions";
 
 const getID = async () => {
-  const token = getPreferenceValues<Pref>().projectAPIID;
+  const token = (await LocalStorage.getItem("journal")) as string;
   return token;
 };
 
@@ -82,8 +82,9 @@ const changeActiveProjectJson = (active: boolean, id: string) => {
 export const QueryDeleteProject = async (project: Project, cache: Cache, notion: Client | undefined) => {
   if (notion === undefined) return;
 
+  const timerAPID = (await LocalStorage.getItem("timer")) as string;
   const qTimer: QueryDatabaseParameters = {
-    database_id: getPreferenceValues<Pref>().timerAPIID,
+    database_id: timerAPID,
     filter: {
       and: [{ property: "Projects", relation: { contains: project.id } }],
     },
@@ -91,23 +92,27 @@ export const QueryDeleteProject = async (project: Project, cache: Cache, notion:
   const resTimers = await notion?.databases.query(qTimer);
   const timers = resTimers === undefined ? [] : getTimers(resTimers);
 
-  const events = await FetchEvents(["all"], getPreferenceValues<Pref>().eventAPIID, true, "", notion);
+  const eventAPIID = (await LocalStorage.getItem("event")) as string;
+  const events = await FetchEvents(["all"], eventAPIID, true, "", notion);
   if (typeof events === "string") {
     showToast({ title: getAPIError(events, "Events"), style: Toast.Style.Failure });
     return;
   }
-  const journals = await FetchJournals(["all"], getPreferenceValues<Pref>().journalAPIID, true, notion);
+  const journalAPIID = (await LocalStorage.getItem("journal")) as string;
+  const journals = await FetchJournals(["all"], journalAPIID, true, notion);
   if (typeof journals === "string") {
     showToast({ title: getAPIError(journals, "Journals"), style: Toast.Style.Failure });
     return;
   }
-  const keystones = await FetchKeystones(["all"], getPreferenceValues<Pref>().keystoneAPIID, true, "", notion);
+  const keystoneAPIID = (await LocalStorage.getItem("keystone")) as string;
+  const keystones = await FetchKeystones(["all"], keystoneAPIID, true, "", notion);
   if (typeof keystones === "string") {
     showToast({ title: getAPIError(keystones, "Keystones"), style: Toast.Style.Failure });
     return;
   }
   const links = project.links;
-  const todos = await FetchedTodos(["all"], getPreferenceValues<Pref>().todoAPIID, true, notion);
+  const todoAPIID = (await LocalStorage.getItem("todo")) as string;
+  const todos = await FetchedTodos(["all"], todoAPIID, true, notion);
   if (typeof todos === "string") {
     showToast({ title: getAPIError(todos, "Todos"), style: Toast.Style.Failure });
     return;