+++ extensions/pilot-ynwr/src/queriesFunctions/TodosQueries.tsx
@@ -1,4 +1,4 @@
-import { Journal, Keystone, Pref, Project, Todo } from "../interfaces/interfaceItems";
+import { Journal, Keystone, Project, Todo } from "../interfaces/interfaceItems";
 import { addJournalJson } from "./JournalsQueries";
 import { APIResponseError, Client } from "@notionhq/client";
 import {
@@ -8,12 +8,12 @@ import {
   UpdatePageParameters,
 } from "@notionhq/client/build/src/api-endpoints";
 import { GetJournalsQuery } from "../fetch/GetQuery";
-import { getPreferenceValues, showToast } from "@raycast/api";
+import { LocalStorage, showToast } from "@raycast/api";
 import { getAPIError } from "../tools/generalTools";
 import { getJournals } from "../fetch/ExportFunctions";
 
 const getToken = async () => {
-  const token = getPreferenceValues<Pref>().todoAPIID;
+  const token = (await LocalStorage.getItem("todo")) as string;
   return token;
 };
 
@@ -30,7 +30,7 @@ export const QueryToogleTodo = async (todo: Todo, notion: Client | undefined) =>
   const bool = !todo.checkbox;
 
   if (bool) {
-    const jouralAPIID = getPreferenceValues<Pref>().journalAPIID;
+    const jouralAPIID = (await LocalStorage.getItem("journal")) as string;
     const q = GetJournalsQuery(true, todo.project.name, false, today);
     const resJournal = await notion?.databases
       .query({ database_id: jouralAPIID as string, ...q } as QueryDatabaseParameters)